src/main.py
import asyncio
import logging
import os
from typing import Dict, Any
from dataclasses import dataclass
from datetime import datetime

import yaml
from pydantic import BaseModel
from swarm import Swarm, Agent

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Configuration
CONFIG_PATH = "config.yaml"

@dataclass
class ContentConfig:
    brand_voice: str
    posting_schedules: Dict[str, Dict[str, str]]
    platform_tokens: Dict[str, str]

@dataclass
class ContentItem:
    id: str
    title: str
    description: str
    source_url: str
    duration: int
    transcript: str
    viral_hooks: list
    variants: list
    thumbnail_path: str
    scheduled_time: datetime

# Load configuration
def load_config(config_path: str) -> ContentConfig:
    with open(config_path, 'r') as f:
        config_data = yaml.safe_load(f)
    return ContentConfig(
        brand_voice=config_data['brand_voice'],
        posting_schedules=config_data['posting_schedules'],
        platform_tokens=config_data['platform_tokens']
    )

# Agent classes

class IngestAgent:
    def __init__(self, config: ContentConfig):
        self.config = config
        self.name = "ingest_agent"
        
    async def process(self, content_url: str) -> ContentItem:
        logger.info(f"Ingesting content from {content_url}")
        
        # Simulate download
        await asyncio.sleep(2)
        
        # Simulate transcription
        transcript = "This is a sample transcript from the video content. It contains key information that will be used to create short-form content."
        
        # Extract viral hooks
        viral_hooks = [
            "This is a viral hook 1",
            "This is a viral hook 2",
            "This is a viral hook 3"
        ]
        
        return ContentItem(
            id=f"content_{datetime.now().timestamp()}",
            title="Sample Video Title",
            description="Sample video description",
            source_url=content_url,
            duration=120,
            transcript=transcript,
            viral_hooks=viral_hooks,
            variants=[],
            thumbnail_path="thumbnail.jpg",
            scheduled_time=datetime.now()
        )

class ClipAgent:
    def __init__(self, config: ContentConfig):
        self.config = config
        self.name = "clip_agent"
        
    async def process(self, content_item: ContentItem) -> ContentItem:
        logger.info(f"Clipping content {content_item.id}")
        
        # Simulate scene detection
        await asyncio.sleep(1)
        
        # Create variants
        variants = [
            {
                "type": "tiktok",
                "duration": 60,
                "start_time": 10,
                "end_time": 70,
                "caption": "This is a TikTok caption"
            },
            {
                "type": "reels",
                "duration": 45,
                "start_time": 20,
                "end_time": 65,
                "caption": "This is a Reels caption"
            },
            {
                "type": "shorts",
                "duration": 30,
                "start_time": 30,
                "end_time": 60,
                "caption": "This is a Shorts caption"
            }
        ]
        
        content_item.variants = variants
        return content_item


class PlatformAgent:
    def __init__(self, config: ContentConfig):
        self.config = config
        self.name = "platform_agent"
        
    async def process(self, content_item: ContentItem) -> ContentItem:
        logger.info(f"Rendering content for platforms {content_item.id}")
        
        # Simulate platform-specific rendering
        await asyncio.sleep(3)
        
        # Generate thumbnails
        thumbnail_path = f"thumbnail_{content_item.id}.jpg"
        
        content_item.thumbnail_path = thumbnail_path
        return content_item


class ScheduleAgent:
    def __init__(self, config: ContentConfig):
        self.config = config
        self.name = "schedule_agent"
        
    async def process(self, content_item: ContentItem) -> ContentItem:
        logger.info(f"Scheduling content {content_item.id}")
        
        # Simulate scheduling
        await asyncio.sleep(1)
        
        # Set scheduled time
        content_item.scheduled_time = datetime.now()
        
        return content_item

# Main coordinator
async def main():
    # Load configuration
    config = load_config(CONFIG_PATH)
    
    # Initialize agents
    ingest_agent = IngestAgent(config)
    clip_agent = ClipAgent(config)
    platform_agent = PlatformAgent(config)
    schedule_agent = ScheduleAgent(config)
    
    # Simulate content ingestion
    content_url = "https://example.com/video.mp4"
    
    # Process through swarm
    content_item = await ingest_agent.process(content_url)
    content_item = await clip_agent.process(content_item)
    content_item = await platform_agent.process(content_item)
    content_item = await schedule_agent.process(content_item)
    
    logger.info(f"Content processing complete: {content_item.id}")
    
    return content_item

if __name__ == "__main__":
    asyncio.run(main())