requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
swarm==0.1.0
pyyaml==6.0.1
asyncio==3.4.3
